refactor: js 확장자 파일을 ts로 변경
